+++ classes/Collision.h
@@ -7,17 +7,17 @@ using namespace std;
 enum collisionDirections {TOPCOLLISION, BOTTOMCOLLISION, LEFTCOLLISION, RIGHTCOLLISION, NOCOLLISION, NULLCOLLISION};
 
 class Collision {
-    public: double x=0, y=0, z=0;
+    public: float x=0, y=0, z=0;
     public: float r=0, g=0, b=0;
-    public: double sizeH=1, sizeV=1;
+    public: float sizeH=1, sizeV=1;
     public: bool isOnPlataform = false;
     public: Collision()= default;
-    public: Collision(double x, double y, double z, float sizeH, float sizeV);
+    public: Collision(float x, float y, float z, float sizeH, float sizeV);
     public: void setSize(float sizeH, float sizeV);
-    public: static collisionDirections checkCollision(map<char, double> mapCollider1, double x1, double y1, map<char, double> mapCollider2, double x2, double y2, bool lastIteration, int* quantityOverLapping);
+    public: static collisionDirections checkCollision(map<char, float> mapCollider1, float x1, float y1, map<char, float> mapCollider2, float x2, float y2, bool lastIteration, int* quantityOverLapping);
 };
 
-Collision :: Collision(double x, double y, double z, float sizeH, float sizeV=-1){
+Collision :: Collision(float x, float y, float z, float sizeH, float sizeV=-1){
     if(sizeV == -1){
         sizeV = sizeH;
     }
@@ -36,7 +36,7 @@ void Collision :: setSize(float updateSizeH, float updateSizeV=-1){
     this->sizeV = updateSizeV;
 }
 
-collisionDirections Collision :: checkCollision(map<char, double> mapCollider1 = map<char, double>(), double x1=0, double y1=0, map<char, double> mapCollider2 = map<char, double>(), double x2=0, double y2=0, bool lastIteration=false, int* quantityOverLapping=nullptr){
+collisionDirections Collision :: checkCollision(map<char, float> mapCollider1 = map<char, float>(), float x1=0, float y1=0, map<char, float> mapCollider2 = map<char, float>(), float x2=0, float y2=0, bool lastIteration=false, int* quantityOverLapping=nullptr){
     collisionDirections typeCollision = NULLCOLLISION;
     bool isOver = mapCollider1['L'] + x1 <= mapCollider2['R'] + x2  &&  mapCollider1['R'] + x1 >= mapCollider2['L'] + x2  &&
                   mapCollider1['B'] + y1 <= mapCollider2['T'] + y2  &&
